Fixed money SaveField bug
The money save field would only be saved into 2 bytes, even though the
save field has type of 4 byte int. This would cause this field to reload
to a different value if a value greater than the 2 byte int max was
input. Fixed by having the money SaveField be 4 bytes.